Refactor utilities from python function to foast into seperate module (preperation PAST)
